refactor(#2615): Change datatypes from class to string
