+++ ui/projects/streampipes/platform-services/src/lib/model/types/semantic-type.service.ts
