+++ ui/src/app/dashboard/sdk/model/datatypes.ts
@@ -16,23 +16,16 @@
  *
  */
 
-import { Vocabulary } from './vocabulary';
-
 export class Datatypes {
-    static readonly Integer = new Datatypes(Vocabulary.XSD, 'integer');
-    static readonly Long = new Datatypes(Vocabulary.XSD, 'long');
-    static readonly Float = new Datatypes(Vocabulary.XSD, 'float');
-    static readonly Boolean = new Datatypes(Vocabulary.XSD, 'boolean');
-    static readonly String = new Datatypes(Vocabulary.XSD, 'string');
-    static readonly Double = new Datatypes(Vocabulary.XSD, 'double');
-    static readonly Number = new Datatypes(Vocabulary.SO, 'Number');
+    static readonly XSD: string = 'http://www.w3.org/2001/XMLSchema#';
+    static readonly SO: string = 'http://schema.org/';
 
-    private constructor(
-        private readonly vocabulary: string,
-        private readonly entity: string,
-    ) {}
+    static readonly Integer = this.XSD + 'integer';
+    static readonly Long = this.XSD + 'long';
+    static readonly Float = this.XSD + 'float';
+    static readonly Boolean = this.XSD + 'boolean';
+    static readonly String = this.XSD + 'string';
+    static readonly Double = this.XSD + 'double';
 
-    toUri() {
-        return this.vocabulary + this.entity;
-    }
+    static readonly Number = this.SO + 'Number';
 }
